RemoveFromGroup {RCy3}	R Documentation

Remove from Group

Description

Removes the specified nodes and edges from the specified group.

Usage

RemoveFromGroup(
  group.name,
  nodes = NULL,
  nodes.by.col = "SUID",
  edges = NULL,
  edges.by.col = "SUID",
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

group.name
Specifies the name used to identify the group

nodes
List of node SUIDs, names, other column values, or keyword: selected, unselected or all. Default is currently selected nodes.

nodes.by.col
name of node table column corresponding to provided nodes list. Default is 'SUID'.

edges
List of edge SUIDs, names, other column values, or keyword: selected, unselected or all. Default is currently selected edges.

edges.by.col
name of edge table column corresponding to provided edges list. Default is 'SUID'.

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

RemoveFromGroup('myGroup')

[Package RCy3 version 2.14.2 Index]