addAnnotationImage {RCy3}	R Documentation

Add Image Annotation

Description

Adds a Image annotation to a Cytoscape network view. The object will also be added to the Annotation Panel in the GUI.

Usage

addAnnotationImage(
  url = NULL,
  x.pos = NULL,
  y.pos = NULL,
  angle = NULL,
  opacity = NULL,
  brightness = NULL,
  contrast = NULL,
  borderThickness = NULL,
  borderColor = NULL,
  borderOpacity = NULL,
  height = NULL,
  width = NULL,
  name = NULL,
  canvas = NULL,
  z.order = NULL,
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

url
URL or path to image file. File paths can be absolute or relative to current working directory. URLs must start with http:// or https://.

x.pos
(optional) X position in pixels from left; default is center of current view

y.pos
(optional) Y position in pixels from top; default is center of current view

angle
(optional) Angle of text orientation; default is 0.0 (horizontal)

opacity
(optional) Opacity of fill color. Must be an integer between 0 and 100; default is 100.

brightness
(optional) Image brightness. Must be an integer between -100 and 100; default is 0

contrast
(optional) Image contrast. Must be an integer between -100 and 100; default is 0

borderThickness
(optional) Integer

borderColor
(optional) Hexidecimal color; default is #000000 (black)

borderOpacity
(optional) Integer between 0 and 100; default is 100.

height
(optional) Height of image; default is based on text height.

width
(optional) Width of image; default is based on text length.

name
(optional) Name of annotation object; default is "Image"

canvas
(optional) Canvas to display annotation, i.e., foreground (default) or background

z.order
(optional) Arrangement order specified by number (larger values are in front of smaller values); default is 0

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

A named list of annotation properties, including UUID

Examples

addAnnotationImage("image.png")
addAnnotationImage("/Users/janedoe/Desktop/image.png", 1000, 1000, name="I2")
addAnnotationImage("https://www.example.com/image.png", 1200, 1000, 30, 
    40, name="I3", canvas="background",z=4)

[Package RCy3 version 2.14.2 Index]