commandEcho {RCy3}	R Documentation

Command Echo

Description

The echo command will display the value of the variable specified by the variableName argument, or all variables if variableName is not provided.

Usage

commandEcho(variable.name = "*", base.url = .defaultBaseUrl)

Arguments

variable.name
(optional) The name of the variable to display. Default is to display all variable values using "*".

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

Value of variable

Examples

commandEcho()

[Package RCy3 version 2.14.2 Index]