commandRunFile {RCy3}	R Documentation

Command Run File

Description

The run command will execute a command script from the file pointed to by the file argument, which should contain Cytoscape commands, one per line. Arguments to the script are provided by the args argument

Usage

commandRunFile(file, args = NULL, base.url = .defaultBaseUrl)

Arguments

file
Path to command script file

args
The script arguments as key:value pairs separated by commas

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

commandRunFile('/path/to/my/file.txt')

[Package RCy3 version 2.14.2 Index]