commandsGET {RCy3}	R Documentation

Commands GET

Description

Using the same syntax as Cytoscape's Command Line Dialog, this function converts a command string into a CyREST query URL, executes a GET request, and parses the result content into an R list object. The function check whether actual call is local or remote first. If remote, requests will go through Jupyter-Bridge.

Usage

commandsGET(cmd.string, base.url = .defaultBaseUrl)

Arguments

cmd.string
(char) command

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

A list, status or None.

Examples

commandsGET('layout get preferred network=current')
commandsGET('network list properties network=current')
commandsGET('layout force-directed defaultNodeMass=1')

[Package RCy3 version 2.14.2 Index]