commandsHelp {RCy3}	R Documentation

Commands Help

Description

Using the same syntax as Cytoscape's Command Line Dialog, this function returns a list of available commands or args. The function check whether actual call is local or remote first. If remote, requests will go through Jupyter-Bridge.

Usage

commandsHelp(cmd.string = "help", base.url = .defaultBaseUrl)

Arguments

cmd.string
(char) command

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

Works with or without 'help' command prefix. Note that if you ask about a command that doesn't have any arguments, this function will run the command!

Value

List of available commands or args

Examples

commandsHelp()
commandsHelp('node')
commandsHelp('node get attribute')

[Package RCy3 version 2.14.2 Index]