createCompositeFilter {RCy3}	R Documentation

Create Composite Filter

Description

Combines filters to control node and edge selection based on previously created filters.

Usage

createCompositeFilter(
  filter.name,
  filter.list,
  type = "ALL",
  hide = FALSE,
  network = NULL,
  base.url = .defaultBaseUrl,
  apply = TRUE
)

Arguments

filter.name
Name for filter.

filter.list
List of filters to combine.

type
(optional) Type of composition, requiring ALL (default) or ANY filters to pass for final node and edge selection.

hide
Whether to hide filtered out nodes and edges. Default is FALSE. Ignored if all nodes or edges are filtered out. This is an alternative to filtering for node and edge selection.

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

apply
(bool) True to execute filter immediately (default); False to define filter but not execute it (available in Cytoscape 3.9+).

Value

List of selected nodes and edges.

Examples

createCompositeFilter("comp1", c("filter1", "filter2"))
createCompositeFilter("comp2", c("filter1", "filter2"), "ANY")
createCompositeFilter("comp3", c("comp1", "filter3"), apply=FALSE)

[Package RCy3 version 2.14.2 Index]