createGroup {RCy3}	R Documentation

Create Group

Description

Create a group from the specified nodes.

Usage

createGroup(
  group.name,
  nodes = NULL,
  nodes.by.col = "SUID",
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

group.name
The name used to identify and optionaly label the group

nodes
List of node SUIDs, names, other column values, or keyword: selected, unselected or all. Default is currently selected nodes.

nodes.by.col
name of node table column corresponding to provided nodes list. Default is 'SUID'.

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

Group SUID

Examples

createGroup('myGroup')

[Package RCy3 version 2.14.2 Index]