cybrowserDialog {RCy3}	R Documentation

Cybrowser Dialog

Description

Launch Cytoscape's internal web browser in a separate window. Provide an id for the window if you want subsequent control of the window e.g., via cybrowser hide.

Usage

cybrowserDialog(
  id = NULL,
  text = NULL,
  title = NULL,
  url = NULL,
  base.url = .defaultBaseUrl
)

Arguments

id
(optional) The identifier for the browser window

text
(optional) HTML text to initially load into the browser

title
(optional) Text to be shown in the title bar of the browser window

url
(optional) The URL the browser should load

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

See Also

cybrowserShow

cybrowserHide

Examples

cybrowserDialog(url='http://cytoscape.org')

[Package RCy3 version 2.14.2 Index]