deleteNetwork {RCy3}	R Documentation

Delete Network

Description

Delete a network from the current Cytoscape session.

Usage

deleteNetwork(network = NULL, base.url = .defaultBaseUrl)

Arguments

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

deleteNetwork()

[Package RCy3 version 2.14.2 Index]