deleteTableColumn {RCy3}	R Documentation

Delete a table column

Description

Delete a column from node, edge or network tables.

Usage

deleteTableColumn(
  column,
  table = "node",
  namespace = "default",
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

column
Name of the column to delete

table
Name of table, e.g., node (default), edge, network

namespace
Namespace of table. Default is "default".

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

A data.frame of column values

Examples

deleteTableColumn('node','group')

[Package RCy3 version 2.14.2 Index]