diffusionAdvanced {RCy3}	R Documentation

Diffusion Advanced

Description

Diffusion will send the selected network view and its selected nodes to a web-based REST service to calculate network propagation. Results are returned and represented by columns in the node table. Advanced operation supports parameters.

Usage

diffusionAdvanced(
  heat.column.name = NULL,
  time = NULL,
  base.url = .defaultBaseUrl
)

Arguments

heat.column.name
(optional) A node column name intended to override the default table column 'diffusion_input'. This represents the query vector and corresponds to h in the diffusion equation.

time
(optional) The extent of spread over the network. This corresponds to t in the diffusion equation.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

Columns are created for each execution of Diffusion and their names are returned in the response. The nodes you would like to use as input should be selected. This will be used to generate the contents of the diffusion_input column, which represents the query vector and corresponds to h in the diffusion equation.

Value

Version number

Examples

diffusionAdvanced()

[Package RCy3 version 2.14.2 Index]