doRequestRemote {RCy3}	R Documentation

doRequestRemote

Description

Do requests remotely by connecting over Jupyter-Bridge.

Usage

doRequestRemote(method, qurl, qbody = NULL, headers = NULL)

Arguments

method
A string to be converted to the REST query namespace

qurl
A named list of values to be converted to REST query parameters

qbody
A named list of values to be converted to JSON

headers
httr headers

Value

httr response

Examples

doRequestRemote()

[Package RCy3 version 2.14.2 Index]