expandGroup {RCy3}	R Documentation

Expand Group

Description

Replaces the group node with member nodes for a set of groups.

Usage

expandGroup(groups = NULL, network = NULL, base.url = .defaultBaseUrl)

Arguments

groups
(optional) List of group names or keywords: all, selected, unselected. Default is the currently selected group.

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

expandGroup()

[Package RCy3 version 2.14.2 Index]