exportFilters {RCy3}	R Documentation

Export Filters

Description

Saves filters to file in JSON format.

Usage

exportFilters(
  filename = "filters.json",
  base.url = .defaultBaseUrl,
  overwriteFile = TRUE
)

Arguments

filename
(char) Full path or path relavtive to current working directory, in addition to the name of the file. Default is "filters.json"

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

overwriteFile
(optional) FALSE allows an error to be generated if the file already exists; TRUE allows Cytoscape to overwrite it without asking. Default value is TRUE.

Details

Unlike other export functions, Cytoscape will automatically overwrite files with the same name. You will not be prompted to confirm or reject overwrite. Use carefully!

Value

None

Examples

exportFilters()

[Package RCy3 version 2.14.2 Index]