exportNetworkToNDEx {RCy3}	R Documentation

Export Network To NDEx

Description

Send a copy of a Cytoscape network to NDEx as a new submission.

Usage

exportNetworkToNDEx(
  username,
  password,
  isPublic,
  network = NULL,
  metadata = NULL,
  ndex.url = "http://ndexbio.org",
  ndex.version = "v2",
  base.url = .defaultBaseUrl
)

Arguments

username
NDEx account username

password
NDEx account password

isPublic
(Boolean) Whether to make the network publicly accessible at NDEx.

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

metadata
(optional) A list of structured information describing the network

ndex.url
(optional) For alternative instances or deployments of NDEx; default is "http://ndexbio.org"

ndex.version
(optional) For alternative versions of the NDEx API; default is "v2"

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

NDEx identifier (externalId) for new submission

Examples

exportNetworkToNDEx("user", "pass", TRUE)

[Package RCy3 version 2.14.2 Index]