exportVisualStyles {RCy3}	R Documentation

Export Visual Styles

Description

Save one or more visual styles to file.

Usage

exportVisualStyles(
  filename = NULL,
  type = "XML",
  styles = NULL,
  base.url = .defaultBaseUrl,
  overwriteFile = TRUE
)

Arguments

filename
(char) Full path or path relavtive to current working directory, in addition to the name of the file. Extension is automatically added based on the type argument. Default is "styles.xml"

type
(optional) Type of data file to export, e.g., XML, JSON (case sensitive). Default is XML. Note: Only XML can be read by importVisualStyles().

styles
(optional) The styles to be exported, listed as a comma-separated string. If no styles are specified, only the current one is exported.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

overwriteFile
(optional) FALSE allows Cytoscape show a message box before overwriting the file if the file already exists; TRUE. allows Cytoscape to overwrite it without asking. Default value is TRUE.

Value

Path to saved file

See Also

importVisualStyles

Examples

exportVisualStyles('/fullpath/myStyle')
exportVisualStyles('/fullpath/myStyle', type = 'JSON')
exportVisualStyles('/fullpath/myStyle', style = 'Minimal,default,Directed')

[Package RCy3 version 2.14.2 Index]