getArrowShapes {RCy3}	R Documentation

Get Arrow Shapes

Description

Retrieve the names of the currently supported 'arrows' – the decorations can (optionally) appear at the ends of edges, adjacent to the nodes they connect, and conveying information about the nature of the nodes' relationship.

Usage

getArrowShapes(base.url = .defaultBaseUrl)

Arguments

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

A list of character strings, e.g., 'DIAMOND', 'T', 'ARROW'

Examples

getArrowShapes()

[Package RCy3 version 2.14.2 Index]