getFirstNeighbors {RCy3}	R Documentation

Get list of nodes neighboring provided list

Description

Returns a non-redundant list of first neighbors of the supplied list of nodes or current node selection.

Usage

getFirstNeighbors(
  node.names = NULL,
  as.nested.list = FALSE,
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

node.names
A list of SUIDs or names from the name column of the node table. Default is currently selected nodes.

as.nested.list
logical Whether to return lists of neighbors per query node

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

A list of unique node names, optionally nested per query node name.

Author(s)

Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon

See Also

selectNodes selectFirstNeighbors

Examples

getFirstNeighbors()

[Package RCy3 version 2.14.2 Index]