getLayoutPropertyValue {RCy3}	R Documentation

Get Layout Property Value

Description

Returns the appropriately typed value of the specified tunable property for the specified layout.

Usage

getLayoutPropertyValue(layout.name, property.name, base.url = .defaultBaseUrl)

Arguments

layout.name
(character) Name of the layout

property.name
(character) Name of the property

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

Run getLayoutNames to list available layouts. Run getLayoutPropertyNames to list properties per layout.

Value

The current value set for this layout property. Typically an integer, numeric or character string value.

Author(s)

Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon

Examples

getLayoutPropertyValue('force-directed','defaultSpringLength')
# 80

[Package RCy3 version 2.14.2 Index]