getLineStyles {RCy3}	R Documentation

Get Line Styles

Description

Retrieve the names of the currently supported line types – values which can be used to render edges, and thus can be used in calls to 'setEdgeLineStyleRule'.

Usage

getLineStyles(base.url = .defaultBaseUrl)

Arguments

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

A list of character strings, e.g., 'SOLID', 'DOT'

Examples

getLineStyles()

[Package RCy3 version 2.14.2 Index]