getNetworkName {RCy3}	R Documentation

Get the name of a network

Description

Retrieve the title of a network

Usage

getNetworkName(suid = NULL, base.url = .defaultBaseUrl)

Arguments

suid
(optional) SUID of the network; default is current network. If a name is provided, then it is validated and returned.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

network name

Examples

getNetworkName()
getNetworkName(1111)

[Package RCy3 version 2.14.2 Index]