getNetworkViewSuid {RCy3}	R Documentation

Get the SUID of a network view

Description

Retrieve the SUID of a network view

Usage

getNetworkViewSuid(network = NULL, base.url = .defaultBaseUrl)

Arguments

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape. If a network view SUID is provided, then it is validated and returned.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

(numeric) Network view suid. The first (presumably only) view associated a network is returned.

Author(s)

Alexander Pico

Examples

getNetworkViewSuid()
getNetworkViewSuid("myNetwork")
# 90

[Package RCy3 version 2.14.2 Index]