getTableColumns {RCy3}	R Documentation

Get table column values

Description

Retrieve one or more columns of data from node, edge or network tables.

Usage

getTableColumns(
  table = "node",
  columns = NULL,
  namespace = "default",
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

table
name of table, e.g., node (default), edge, network

columns
names of columns to retrieve values from as list object or comma-separated list; default is all columns

namespace
namespace of table; default is "default"

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

The 'SUID' column is always retrieved along with specified columns. The 'SUID' values are used as row.names in the returned data.frame.

Value

A data.frame of column values

Examples

getTableColumns()
getTableColumns('node','group')

[Package RCy3 version 2.14.2 Index]