getTableValue {RCy3}	R Documentation

Get table cell value

Description

Retrieve the value from a specific row and column from node, edge or network tables.

Usage

getTableValue(
  table,
  row.name,
  column,
  namespace = "default",
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

table
name of table, e.g., node, edge, network

row.name
Node, edge or network name, i.e., the value in the "name" column

column
name of column to retrieve values from

namespace
namespace of table; default is "default"

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

A data.frame of column values

Examples

getTableValue('node','node 1', 'score')

[Package RCy3 version 2.14.2 Index]