hideSelectedEdges {RCy3}	R Documentation

Hide Selected Edges

Description

Hide (but do not delete) the currently selected edges, by setting the Visible property bypass value to false.

Usage

hideSelectedEdges(network = NULL, base.url = .defaultBaseUrl)

Arguments

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

This method permanently overrides any default values or mappings defined for this visual property of the edge or edges specified. This method ultimately calls the generic function, setEdgePropertyBypass, which can be used to set any visual property. To restore defaults and mappings, use unhideEdges or unhideAll.

Value

None

See Also

setEdgePropertyBypass, hideEdges, unhideEdges, unhideAll

Examples

hideSelectedEdges()

[Package RCy3 version 2.14.2 Index]