loadTableData {RCy3}	R Documentation

Loads data into Cytoscape tables keyed by row

Description

This function loads data into Cytoscape node/edge/network tables provided a common key, e.g., name. Data.frame column names will be used to set Cytoscape table column names.

Usage

loadTableData(
  data,
  data.key.column = "row.names",
  table = "node",
  table.key.column = "name",
  namespace = "default",
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

data
(data.frame) each row is a node and columns contain node attributes

data.key.column
(char) name of data.frame column to use as key; default is "row.names"

table
(char) name of Cytoscape table to load data into, e.g., node, edge or network; default is "node"

table.key.column
(char) name of Cytoscape table column to use as key; default is "name"

namespace
namespace of table, e.g., default

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

Numeric values will be stored as Doubles in Cytoscape tables. Integer values will be stored as Integers. Character or mixed values will be stored as Strings. Logical values will be stored as Boolean. Lists are stored as Lists by CyREST v3.9+. Existing columns with the same names will keep original type but values will be overwritten.

Value

server response

[Package RCy3 version 2.14.2 Index]