makeSimpleGraph {RCy3}	R Documentation

Make Simple Graph

Description

This function creates a simple graphNEL object with various node and edge attribute types to help demonstrate round trip conversion with Cytoscape networks via RCy3.

Usage

makeSimpleGraph()

Value

A graphNEL object with a few nodes, edges and attributes

See Also

createNetworkFromGraph, createGraphFromNetwork, makeSimpleIgraph

Examples

{
makeSimpleGraph()
}

[Package RCy3 version 2.14.2 Index]