makeSimpleIgraph {RCy3}	R Documentation

Make Simple Igraph

Description

This function creates a simple iGraph object with various node and edge attribute types to help demonstrate round trip conversion with Cytoscape networks via RCy3.

Usage

makeSimpleIgraph()

Value

A igraph object with a few nodes, edges and attributes

See Also

createNetworkFromIgraph, createIgraphFromNetwork, makeSimpleGraph

Examples

{
makeSimpleIgraph()
}

[Package RCy3 version 2.14.2 Index]