openSession {RCy3}	R Documentation

Open Session File or URL

Description

Open a session file or URL. This will clear all networks, tables and styles associated with current session. Be sure to saveSession first.

Usage

openSession(file.location = NULL, base.url = .defaultBaseUrl)

Arguments

file.location
File path or URL (with 'http' or 'https' prefix). Default is a sample session file.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

See Also

saveSession

Examples

openSession('/fullpath/mySession.CYS')

[Package RCy3 version 2.14.2 Index]