sandboxSet {RCy3}	R Documentation

sandboxSet

Description

Set a new default sandbox, creating it if necessary. A sandbox is the root for the file system used for all file operations. When running standalone on the same workstation as Cytoscape, the default sandbox is the directory that's current for the R kernel. When running in a Notebook or remote server, the default sandbox is the 'default_sandbox' created automatically under the under the filetransfer directory in the CytoscapeConfiguration directory. Naming a sandbox with this function creates a new sub-directory as a sibling to 'default_sandbox' and uses it for subsequent file operations. Setting a None sandbox uses the default sandbox instead. Sandboxes are highly recommended as an aid to creating workflows that can be shared with others.

Usage

sandboxSet(
  sandboxName,
  copySamples = TRUE,
  reinitialize = TRUE,
  base.url = .defaultBaseUrl
)

Arguments

sandboxName
Name of new default sandbox. None means to use the original default sandbox

copySamples
True to copy the Cytoscape sampleData into the sandbox

reinitialize
True to delete sandbox contents (if any) if sandbox already exists

base.url
Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://127.0.0.1:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

sandbox path in Cytoscape workstation's file system

Examples

sandboxSet()

[Package RCy3 version 2.14.2 Index]