saveSession {RCy3}	R Documentation

Save Session to File

Description

Saves the current Cytoscape session as a CYS file.

Usage

saveSession(filename = NULL, base.url = .defaultBaseUrl, overwriteFile = TRUE)

Arguments

filename
Full path or path relavtive to current working directory, in addition to the name of the file. The .cys extension is automatically added. Leave blank to update previously saved session file.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

overwriteFile
(optional) FALSE allows an error to be generated if the file already exists; TRUE allows Cytoscape to overwrite it without asking. Default value is TRUE.

Details

If no filename is provided, then it attempts to save to an existing CYS file associated with the session. If filename already exists, then it is overwritten.

Unlike most export functions in RCy3, Cytoscape will automatically overwrite CYS session files with the same name. You will not be prompted to confirm or reject overwrite. Use carefully!

Value

server response

Examples

saveSession('/fullpath/mySession')
saveSession() 

[Package RCy3 version 2.14.2 Index]