selectAllNodes {RCy3}	R Documentation

Select all nodes

Description

Selects all nodes in a Cytoscape Network

Usage

selectAllNodes(network = NULL, base.url = .defaultBaseUrl)

Arguments

network
(optional) Name or SUID of the network into which you want to select; default is "current" network

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

Selects all nodes in a specified network.

Author(s)

Alexander Pico, Julia Gustavsen

See Also

selectNodes

Examples

selectAllNodes()

[Package RCy3 version 2.14.2 Index]