selectEdges {RCy3}	R Documentation

Select Edges

Description

Select edges in the network by SUID, name or other column values.

Usage

selectEdges(
  edges,
  by.col = "SUID",
  preserve.current.selection = TRUE,
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

edges
List of edge SUIDs, names or other column values

by.col
Edge table column to lookup up provide edge values. Default is 'SUID'.

preserve.current.selection
boolean Whether to maintain previously selected edges.

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

list of newly selected edge SUIDs

Author(s)

AlexanderPico, Tanja Muetze, Georgi Kolishovski, Paul Shannon

Examples

selectEdges()

[Package RCy3 version 2.14.2 Index]