setCatchupFilterSecs {RCy3}	R Documentation

Set Catchup Filter Delay

Description

This function sets an internal delay variable that allows Cytoscape to "catchup" prior to subsequent functions. Call without specifying secs to restore default value.

Usage

setCatchupFilterSecs(secs = 1)

Arguments

secs
Number of seconds to delay.

Details

This delay is only necessary while concurrency bugs exist in the Cytoscape application. This delay may need to be increased from the default value in certain use cases, e.g., larger networks.

Value

None

See Also

setModelPropagationSecs, setCatchupNetworkSecs

Examples

{
setCatchupFilterSecs(2)
setCatchupFilterSecs() #restores default delay
}

[Package RCy3 version 2.14.2 Index]