setDefaultSandbox {RCy3}	R Documentation

setDefaultSandbox

Description

Set and return the sandbox properties to be used as a default, probably based on whether a Notebook is running.

Usage

setDefaultSandbox(newSandbox = NULL, ...)

Arguments

newSandbox
newSandbox

...
...

Value

default sandbox

Examples

setDefaultSandbox()

[Package RCy3 version 2.14.2 Index]