Set Edge Font Face Default

Set the default edge font.

setEdgeFontFaceDefault(
  new.font,
  style.name = "default",
  base.url = .defaultBaseUrl
)

Arguments

new.font
String specification of font face, style and size, e.g., "SansSerif,plain,12" or "Dialog,plain,10"

style.name
Name of style; default is "default" style

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

# \donttest{
setEdgeFontFaceDefault("Dialog,plain,10")
# }