setEdgeFontFaceMapping {RCy3}	R Documentation

Set Edge Font Face Mapping

Description

Sets font face for edge labels.

Usage

setEdgeFontFaceMapping(
  table.column,
  table.column.values,
  fonts,
  mapping.type = "d",
  default.font = NULL,
  style.name = NULL,
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

table.column
Name of Cytoscape table column to map values from

table.column.values
List of values from Cytoscape table to be used in mapping

fonts
List of string specifications of font face, style and size, e.g., c("SansSerif,plain,12", "Dialog,plain,10")

mapping.type
(char) discrete or passthrough (d,p); default is discrete

default.font
String specification of font face, style and size, e.g., "SansSerif,plain,12" or "Dialog,plain,10"

style.name
Name of style; default is "default" style

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

setEdgeFontFaceMapping("myfonts", c("normal","small"),
c("SansSerif,plain,12", "Dialog,plain,10"))

[Package RCy3 version 2.14.2 Index]