setEdgeOpacityDefault {RCy3}	R Documentation

Set Edge Opacity Default

Description

Set default opacity value for all unmapped edges.

Usage

setEdgeOpacityDefault(
  new.opacity,
  style.name = NULL,
  base.url = .defaultBaseUrl
)

Arguments

new.opacity
Numeric values between 0 and 255; 0 is invisible.

style.name
Name of style; default is "default" style.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

setEdgeOpacityDefault(50)

[Package RCy3 version 2.14.2 Index]