setEdgeTargetArrowShapeDefault {RCy3}	R Documentation

Set Edge Target Arrow Shape Default

Description

Set the default edge target arrow shape.

Usage

setEdgeTargetArrowShapeDefault(
  new.shape,
  style.name = NULL,
  base.url = .defaultBaseUrl
)

Arguments

new.shape
Name of shape, e.g., ARROW, T, etc (see getArrowShapes)

style.name
Name of style; default is "default" style

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

setEdgeTargetArrowShapeDefault('ARROW')

[Package RCy3 version 2.14.2 Index]