setLayoutProperties {RCy3}	R Documentation

Set Layout Properties

Description

Sets the specified properties for the specified layout. Unmentioned properties are left unchanged.

Usage

setLayoutProperties(layout.name, properties.list, base.url = .defaultBaseUrl)

Arguments

layout.name
(character) Name of the layout

properties.list
(list) List of one or more property=value pairs

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

Run getLayoutNames to list available layouts. Run getLayoutPropertyNames to list properties per layout.

Value

None

Author(s)

Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon

Examples

setLayoutProperties('force-directed', list(defaultSpringLength=50, defaultSpringCoefficient=6E-04))
# Successfully updated the property 'defaultSpringLength'.
# Successfully updated the property 'defaultSpringCoefficient'.

[Package RCy3 version 2.14.2 Index]