setNetworkPropertyBypass {RCy3}	R Documentation

Set Network Property Bypass

Description

Set bypass values for any network property, overriding default values defined by any visual style.

Usage

setNetworkPropertyBypass(
  new.value,
  visual.property,
  bypass = TRUE,
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

new.value
Value to set

visual.property
Name of a visual property. See getVisualPropertyNames.

bypass
Whether to set permanent bypass value. Default is TRUE.

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

This method permanently overrides any default values or mappings defined for the visual properties of the node or nodes specified. To restore defaults and mappings, use clearNodePropertyBypass.

Value

None

See Also

clearNetworkPropertyBypass

Examples

setNetworkPropertyBypass()

[Package RCy3 version 2.14.2 Index]