Set Node Border Width Bypass

Override the border width for particular nodes.

setNodeBorderWidthBypass(
  node.names,
  new.sizes,
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

node.names
List of node names

new.sizes
List of size values, or single value

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Details

This method permanently overrides any default values or mappings defined for this visual property of the node or nodes specified. This method ultimately calls the generic function, setNodePropertyBypass, which can be used to set any visual property. To restore defaults and mappings, use clearNodePropertyBypass, see examples.

See also

Examples

# \donttest{
setNodeBorderWidthBypass('Node 1', 5)
setNodeBorderWidthBypass(c('Node 1','Node 2'), 5)
clearNodePropertyBypass(c('Node 1','Node 2'), 'NODE_BORDER_WIDTH')
# }