setNodeComboOpacityMapping {RCy3}	R Documentation

Set Node Combo Opacity Mapping

Description

Sets opacity for node fill, border and label all together.

Usage

setNodeComboOpacityMapping(
  table.column,
  table.column.values = NULL,
  opacities = NULL,
  mapping.type = "c",
  default.opacity = NULL,
  style.name = NULL,
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

table.column
Name of Cytoscape table column to map values from

table.column.values
List of values from Cytoscape table to be used in mapping. Leave NULL to perform an automatic mapping to all column values.

opacities
(integer) values between 0 and 255; 0 is invisible. A range of 50 to 255 is used by default for automatic mapping.

mapping.type
(char) continuous, discrete or passthrough (c,d,p); default is continuous

default.opacity
Opacity value to set as default for all unmapped values

style.name
Name of style; default is "default" style

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

setNodeComboOpacityMapping('score')
setNodeComboOpacityMapping('score', opacities=c(0,100))
setNodeComboOpacityMapping('score', c(-5,5), c(50,255))

[Package RCy3 version 2.14.2 Index]