setNodeCustomPieChart {RCy3}	R Documentation

Set Node Custom Pie Chart

Description

Makes a pie chart per node using specified node table columns by setting a default custom graphic style.

Usage

setNodeCustomPieChart(
  columns,
  colors = NULL,
  startAngle = 0,
  slot = 1,
  style.name = NULL,
  base.url = .defaultBaseUrl
)

Arguments

columns
List of node column names to be displayed.

colors
(optional) List of colors to be matched with columns or with range, depending on type. Default is a set of colors from an appropriate Brewer palette.

startAngle
(optional) Angle to start filling pie. Default is 0.0.

slot
(optional) Which custom graphics slot to modify. Slots 1-9 are available for independent charts, gradients and images. Default is 1.

style.name
(optional) Name of style; default is "default" style

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

See Also

setNodeCustomPosition, removeNodeCustomGraphics

Examples

setNodeCustomPieChart(c("data1","data2","data3"))

[Package RCy3 version 2.14.2 Index]