setNodeCustomPosition {RCy3}	R Documentation

Set Node Custom Position

Description

Adjust the position of a custom graphic relative to its node.

Usage

setNodeCustomPosition(
  nodeAnchor = "C",
  graphicAnchor = "C",
  justification = "c",
  xOffset = 0,
  yOffset = 0,
  slot = 1,
  style.name = NULL,
  base.url = .defaultBaseUrl
)

Arguments

nodeAnchor
Position on node to place the graphic: NW,N,NE,E,SE,S,SW,W or C for center (default)

graphicAnchor
Position on graphic to place on node: NW,N,NE,E,SE,S,SW,W or C for center (default)

justification
Positioning of content within graphic: l,r,c (default)

xOffset
Additional offset in the x direction

yOffset
Additional offset in the y direction

slot
(optional) Which custom graphics slot to modify. Slots 1-9 are available for independent charts, gradients and images. Default is 1.

style.name
(optional) Name of style; default is "default" style

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

setNodeCustomPosition()

[Package RCy3 version 2.14.2 Index]