setNodeCustomRadialGradient {RCy3}	R Documentation

Set Node Custom Radial Gradient

Description

Makes a gradient fill per node by setting a default custom graphic style.

Usage

setNodeCustomRadialGradient(
  colors = c("#DDDDDD", "#888888"),
  anchors = c(0, 1),
  xCenter = 0.5,
  yCenter = 0.5,
  slot = 1,
  style.name = NULL,
  base.url = .defaultBaseUrl
)

Arguments

colors
(optional) List of colors to define gradient

anchors
(optional) Position of colors from 0.0 to 1.0.

xCenter
(optional) X position for center of radial effect from 0.0 to 1.0. Default is 0.5.

yCenter
(optional) Y position for center of radial effect from 0.0 to 1.0. Default is 0.5.

slot
(optional) Which custom graphics slot to modify. Slots 1-9 are available for independent charts, gradients and images. Default is 1.

style.name
(optional) Name of style; default is "default" style

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

setNodeCustomRadialGradient()

[Package RCy3 version 2.14.2 Index]