setNodeFillOpacityBypass {RCy3}	R Documentation

Set Node Fill Opacity Bypass

Description

Override the fill opacity for particular nodes.

Usage

setNodeFillOpacityBypass(
  node.names,
  new.values,
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

node.names
List of node names or SUIDs

new.values
List of values to set, or single value

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

This method permanently overrides any default values or mappings defined for this visual property of the node or nodes specified. This method ultimately calls the generic function, setNodePropertyBypass, which can be used to set any visual property. To restore defaults and mappings, use clearNodePropertyBypass, see examples.

Value

None

See Also

setNodePropertyBypass, clearNodePropertyBypass

Examples

setNodeFillOpacityBypass('Node 1', 100)
setNodeFillOpacityBypass(c('Node 1','Node 2'), 100)
clearNodePropertyBypass(c('Node 1','Node 2'), 'NODE_TRANSPARENCY')

[Package RCy3 version 2.14.2 Index]