setNodeLabelColorDefault {RCy3}	R Documentation

Set Node Label Color Default

Description

Set the default node label color.

Usage

setNodeLabelColorDefault(
  new.color,
  style.name = NULL,
  base.url = .defaultBaseUrl
)

Arguments

new.color
Color as hex code, e.g., #FD5903

style.name
Name of style; default is "default" style

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

setNodeLabelColorDefault('#FD5903')

[Package RCy3 version 2.14.2 Index]