setNodeTooltipMapping {RCy3}	R Documentation

Set Node Tooltip Mapping

Description

Pass the values from a table column to display as node tooltips.

Usage

setNodeTooltipMapping(
  table.column,
  style.name = NULL,
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

table.column
Name of Cytoscape table column to map values from

style.name
Name of style; default is "default" style

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

setNodeTooltipMapping('description')

[Package RCy3 version 2.14.2 Index]