unhideAll {RCy3}	R Documentation

Unhide All

Description

Unhide all previously hidden nodes and edges, by clearing the Visible property bypass value.

Usage

unhideAll(network = NULL, base.url = .defaultBaseUrl)

Arguments

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

This method sets node and edge visibility bypass to true, overriding any defaults or mappings. Pending CyREST updates, this method will ultimately call the generic function, clearEdgePropertyBypass, which can be used to clear any visual property.

Value

None

See Also

clearEdgePropertyBypass, unhideNodes unhideEdges

Examples

unhideAll()

[Package RCy3 version 2.14.2 Index]