uninstallApp {RCy3}	R Documentation

Uninstall App

Description

Uninstall an app from Cytoscape.

Usage

uninstallApp(app, base.url = .defaultBaseUrl)

Arguments

app
Name of app

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

None

Examples

uninstallApp()

[Package RCy3 version 2.14.2 Index]