updateGroupAnnotation {RCy3}	R Documentation

Update Group Annotation

Description

Updates a group annotation, changing the given properties.

Usage

updateGroupAnnotation(
  name = NULL,
  annotationName = NULL,
  x.pos = NULL,
  y.pos = NULL,
  angle = NULL,
  canvas = NULL,
  z.order = NULL,
  network = NULL,
  base.url = .defaultBaseUrl
)

Arguments

name
(optional) Name of annotation object

annotationName
Name of annotation by UUID or Name

x.pos
(optional) X position in pixels from left; default is center of current view

y.pos
(optional) Y position in pixels from top; default is center of current view

angle
(optional) Angle of text orientation; default is 0.0 (horizontal)

canvas
(optional) Canvas to display annotation, i.e., foreground (default) or background

z.order
(optional) Arrangement order specified by number (larger values are in front of smaller values); default is 0

network
(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Value

A named list of annotation properties, including UUID

Examples

updateGroupAnnotation("test1", "annotationName")

[Package RCy3 version 2.14.2 Index]