updateStyleMapping {RCy3}	R Documentation

Updates a visual property mapping in a style

Description

Updates the visual property mapping, overriding any prior mapping. Creates a visual property mapping if it doesn't already exist in the style.

Usage

updateStyleMapping(style.name, mapping, base.url = .defaultBaseUrl)

Arguments

style.name
(char) name for style

mapping
a single visual property mapping, see mapVisualProperty

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

Requires visual property mappings to be previously created, see mapVisualProperty.

Value

server response

See Also

mapVisualProperty

Examples

updateStyleMapping('myStyle',mapVisualProperty('node label','name','p'))

[Package RCy3 version 2.14.2 Index]