layoutCopycat {RCy3}	R Documentation

Copy a layout from one network to another

Description

Sets the coordinates for each node in the target network to the coordinates of a matching node in the source network.

Usage

layoutCopycat(
  sourceNetwork,
  targetNetwork,
  sourceColumn = "name",
  targetColumn = "name",
  gridUnmapped = TRUE,
  selectUnmapped = TRUE,
  base.url = .defaultBaseUrl
)

Arguments

sourceNetwork
(character) The name of network to get node coordinates from

targetNetwork
(character) The name of the network to apply coordinates to

sourceColumn
(optional character) The name of column in the sourceNetwork node table used to match nodes; default is 'name'

targetColumn
(optional character) The name of column in the targetNetwork node table used to match nodes; default is 'name'

gridUnmapped
(optional character) If this is set to true, any nodes in the target network that could not be matched to a node in the source network will be laid out in a grid; default is TRUE

selectUnmapped
optional character) If this is set to true, any nodes in the target network that could not be matched to a node in the source network will be selected in the target network; default is TRUE

base.url
(optional) Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://localhost:1234 and the latest version of the CyREST API supported by this version of RCy3.

Details

Optional parameters such as gridUnmapped and selectUnmapped determine the behavior of target network nodes that could not be matched.

Value

None

Examples

layoutCopycat('network1','network2')

[Package RCy3 version 2.14.2 Index]